requirements.txt
PyMuPDF
google-adk[database] == 0.3.0
psutil == 5.9.5
litellm ==1.66.3
google-generativeai == 0.8.5
python-dotenv == 1.1.0

